create shared auth utility and move shared directory into repo instead of pointing to submodule repo
