--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "shared"]
-	path = shared
-	url = https://github.com/Code-4-Community/shared.git
