+++ apps/backend/lambdas/expenditures/package.json
@@ -23,6 +23,7 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
+    "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
     "aws-jwt-verify": "^5.1.1",
     "aws-lambda": "^1.0.7",
     "jest": "^30.2.0",
