+++ apps/backend/lambdas/projects/auth.ts
@@ -1,144 +1,19 @@
-import { CognitoJwtVerifier } from 'aws-jwt-verify';
+import { authenticateRequest as _authenticateRequest } from '@branch/lambda-auth';
 import db from './db';
 
-const COGNITO_USER_POOL_ID = process.env.COGNITO_USER_POOL_ID!;
-const COGNITO_CLIENT_ID = process.env.COGNITO_APP_CLIENT_ID!;
+export * from '@branch/lambda-auth';
 
-// Create verifier instance lazily (only when needed)
-let verifier: any = null;
-
-function getVerifier() {
-  if (!verifier) {
-    if (!COGNITO_USER_POOL_ID) {
-      throw new Error('COGNITO_USER_POOL_ID environment variable is not set');
-    }
-    verifier = CognitoJwtVerifier.create({
-      userPoolId: COGNITO_USER_POOL_ID,
-      tokenUse: 'access',
-      clientId: COGNITO_CLIENT_ID,
-    });
-  }
-  return verifier;
-}
-
-export interface AuthenticatedUser {
-  cognitoSub: string;
-  userId?: number;
-  email: string;
-  isAdmin?: boolean;
-  cognitoGroups?: string[];
-}
-
-export interface AuthContext {
-  user?: AuthenticatedUser;
-  isAuthenticated: boolean;
-}
-
-/**
- * Encode a JWT token (for testing purposes)
- * Creates a mock JWT with the standard three-part format
- */
-export function encodeJWT(payload: any): string {
-  const header = Buffer.from(JSON.stringify({ alg: 'none', typ: 'JWT' })).toString('base64');
-  const body = Buffer.from(JSON.stringify(payload)).toString('base64');
-  const signature = '';
-  
-  // Format: header.payload.signature
-  return `${header}.${body}.${signature}`;
-}
-
-/**
- * Extract and validate JWT token from Authorization header
- * Format: "Bearer <token>"
- */
-export function extractTokenFromHeader(authorizationHeader: string | undefined): string | null {
-  if (!authorizationHeader) return null;
-  
-  const parts = authorizationHeader.split(' ');
-  if (parts.length !== 2 || parts[0].toLowerCase() !== 'bearer') {
-    return null;
-  }
-  
-  return parts[1];
-}
-
-/**
- * Decode JWT token without verification (for development/testing)
- * In production, you would verify with Cognito public keys
- */
-export function decodeJWT(token: string): any {
-  try {
-    const parts = token.split('.');
-    if (parts.length !== 3) return null;
-    
-    // Decode payload (second part)
-    const payload = parts[1];
-    const decoded = Buffer.from(payload, 'base64').toString('utf8');
-    return JSON.parse(decoded);
-  } catch (error) {
-    console.error('Error decoding JWT:', error);
-    return null;
-  }
-}
-
-/**
- * Authenticate request and return user info
- * This validates the token and fetches user details from database
- */
 export async function authenticateRequest(
-  authorizationHeader: string | undefined
-): Promise<{ user: AuthenticatedUser; error?: string }> {
-  const token = extractTokenFromHeader(authorizationHeader);
-  
-  if (!token) {
-    return { user: null as any, error: 'Missing or invalid Authorization header' };
-  }
-
-  const decoded = decodeJWT(token);
-  
-  if (!decoded || !decoded.sub) {
-    return { user: null as any, error: 'Invalid token' };
-  }
-
-  // The 'sub' claim contains the Cognito user ID
-  const cognitoSub = decoded.sub;
-  const email = decoded.email;
-
-  try {
-    // Fetch user from database
-    const dbUser = await db
-      .selectFrom('branch.users')
-      .where('cognito_sub', '=', cognitoSub)
-      .selectAll()
-      .executeTakeFirst();
-
-    if (!dbUser) {
-      return { user: null as any, error: 'User not found in database' };
-    }
-
-    const user: AuthenticatedUser = {
-      cognitoSub,
-      email: email || dbUser.email,
-      userId: dbUser.user_id,
-      isAdmin: dbUser.is_admin || false,
-    };
-
-    return { user };
-  } catch (error) {
-    console.error('Database lookup error:', error);
-    return { user: null as any, error: 'Failed to authenticate user' };
-  }
+  event: any,
+): Promise<import('@branch/lambda-auth').AuthContext> {
+  return _authenticateRequest(db, event);
 }
 
-/**
- * Check if user has access to a project
- * Access granted if:
- * - User is admin, OR
- * - User is a member of the project
- */
-export async function canAccessProject(userId: number, projectId: number): Promise<boolean> {
+export async function canAccessProject(
+  userId: number,
+  projectId: number,
+): Promise<boolean> {
   try {
-    // Check if user is admin
     const user = await db
       .selectFrom('branch.users')
       .where('user_id', '=', userId)
@@ -147,7 +22,6 @@ export async function canAccessProject(userId: number, projectId: number): Promi
 
     if (user?.is_admin) return true;
 
-    // Check if user is a member of the project
     const membership = await db
       .selectFrom('branch.project_memberships')
       .where('user_id', '=', userId)
@@ -162,15 +36,11 @@ export async function canAccessProject(userId: number, projectId: number): Promi
   }
 }
 
-/**
- * Check if user has edit access to a project
- * Access granted if:
- * - User is admin, OR
- * - User is a PI, Accountant, or Admin in the project
- */
-export async function canEditProject(userId: number, projectId: number): Promise<boolean> {
+export async function canEditProject(
+  userId: number,
+  projectId: number,
+): Promise<boolean> {
   try {
-    // Check if user is admin
     const user = await db
       .selectFrom('branch.users')
       .where('user_id', '=', userId)
@@ -179,7 +49,6 @@ export async function canEditProject(userId: number, projectId: number): Promise
 
     if (user?.is_admin) return true;
 
-    // Check if user has edit role in project
     const membership = await db
       .selectFrom('branch.project_memberships')
       .where('user_id', '=', userId)
@@ -189,7 +58,6 @@ export async function canEditProject(userId: number, projectId: number): Promise
 
     if (!membership) return false;
 
-    // PI, Accountant, and Admin can edit
     const editableRoles = ['PI', 'Accountant', 'Admin'];
     return editableRoles.includes(membership.role);
   } catch (error) {
@@ -198,11 +66,6 @@ export async function canEditProject(userId: number, projectId: number): Promise
   }
 }
 
-/**
- * Check if user can create projects
- * Access granted if:
- * - User is admin
- */
 export async function canCreateProject(userId: number): Promise<boolean> {
   try {
     const user = await db
