+++ apps/backend/lambdas/projects/package.json
@@ -21,6 +21,7 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
+    "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
     "aws-jwt-verify": "^5.1.1",
     "jest": "^30.2.0",
     "kysely": "^0.28.8",
