+++ apps/backend/lambdas/reports/Dockerfile
@@ -1,22 +1,19 @@
 FROM node:20-alpine
 
+# Build shared auth package (file:../../../../shared/lambda-auth from /app resolves to /shared/lambda-auth)
+WORKDIR /shared/lambda-auth
+COPY shared/lambda-auth/package.json shared/lambda-auth/tsconfig.json ./
+COPY shared/lambda-auth/src ./src/
+RUN npm install && npm run build
+
 WORKDIR /app
+COPY apps/backend/lambdas/reports/package*.json ./
+RUN npm install --no-package-lock
+COPY apps/backend/lambdas/reports/ .
 
-# Copy package files
-COPY package*.json ./
-
-# Install dependencies
-RUN npm install
-
-# Copy source files
-COPY . .
-
-# Expose port
 EXPOSE 3000
 
-# Health check
 HEALTHCHECK --interval=30s --timeout=3s --start-period=5s --retries=3 \
   CMD wget --no-verbose --tries=1 --spider http://localhost:3000/reports/health || exit 1
 
-# Start the dev server
 CMD ["npm", "run", "dev"]
