+++ apps/backend/lambdas/reports/package.json
@@ -24,6 +24,7 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
+    "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
     "@aws-sdk/client-s3": "^3.995.0",
     "aws-jwt-verify": "^5.1.1",
     "aws-lambda": "^1.0.7",
