+++ shared/lambda-auth/src/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export { extractToken, authenticateRequest } from './authenticate';
+export { checkAuthorization } from './authorize';
